Add -strict flag and ignore message errors by default.
By default, copy the rest of the message to stdout and exit
with status code 0 when a non-I/O error is encountered.
These errors currently include malformed header fields, a
missing blank line between the header and body, and a
missing MIME boundary delimiter.

A new -strict flag can be passed to get the old behavior of
exiting with status code 1 in response to bad messages.